+++ src/entries/wvui.ts
@@ -2,8 +2,12 @@ import WvuiButton from '@/components/button/Button.vue';
 import WvuiInput from '@/components/input/Input.vue';
 import WvuiTypeaheadSuggestion from '@/components/typeahead-suggestion/TypeaheadSuggestion.vue';
 
+// Export version as a named export so that the default export so that the default export can be
+// passed to the Vue app instance's components directly.
+export const version = VERSION;
+
+// Export all components available in the library.
 export default {
-	version: VERSION,
 	WvuiButton,
 	WvuiInput,
 	WvuiTypeaheadSuggestion