+++ changelog.md
@@ -4,6 +4,7 @@ Versions and bullets are arranged chronologically from latest to oldest.
 
 ## v0.0.2 (unreleased)
 
+-   [dev] Add npm script to clear Jest cache
 -   [build][dev] Add production and pre-release scripts
 
 ## v0.0.1